Move components to another folder
